// web-client/src/api/api.ts
export interface ActivityLookupDto {
  id?: string;
  description?: string | undefined;
  tags?: Tag[] | undefined;
  project?: Project;
  startInMilliseconds?: number;
  endInMilliseconds?: number | undefined;
}

export interface CreateActivityDto {
  description?: string | undefined;
  tagIds?: string[] | undefined;
  projectId?: string | undefined;
  startInMilliseconds?: number;
  endInMilliseconds?: number | undefined;
}

export interface CreateProjectDto {
  title?: string | undefined;
}

export interface CreateTagDto {
  title?: string | undefined;
}

export interface Project {
  id?: string;
  creationDate?: Date;
  editDate?: Date | undefined;
  userId?: string;
  title?: string | undefined;
}

export interface ProjectDetailVm {
  id?: string;
  title?: string | undefined;
  creationDate?: Date;
  editDate?: Date | undefined;
}

export interface ProjectListVm {
  projects?: ProjectLookupDto[] | undefined;
}

export interface ProjectLookupDto {
  id?: string;
  title?: string | undefined;
}

export interface StartActivityDto {
  description?: string | undefined;
  tagIds?: string[] | undefined;
  projectId?: string | undefined;
}

export interface Tag {
  id?: string;
  creationDate?: Date;
  editDate?: Date | undefined;
  userId?: string;
  title?: string | undefined;
}

export interface TagDetailVm {
  id?: string;
  title?: string | undefined;
  creationDate?: Date;
  editDate?: Date | undefined;
}

export interface TagListVm {
  tags?: TagLookupDto[] | undefined;
}

export interface TagLookupDto {
  id?: string;
  title?: string | undefined;
}